requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
python-multipart
jinja2
python-jose[cryptography]
passlib[bcrypt]
starlette
itsdangerous
